+++ lib/src/dartdap/core/rdn.dart
@@ -1,0 +1,2 @@
+/// See https://ldap.com/ldap-dns-and-rdns/ for rules for DNs / RDNs
+///